Revert "Plotter refactoring, new features and performance work"
Revert "bookmarks: restore Delete key"
Revert "Use a QImage for the waterfall"
Revert "plotter: clear waterfall using Delete key"
Revert "Invalidate max, min, & hist before updateOverlay & redraw"
Revert "Fix the pixels/bin calculation"
Revert "Fix drawing of black areas in waterfall"
Revert "Avoid resetting histogram, min & max unnecessarily"
Revert "Add transparency to plotter overlay elements"
Revert "Don't run FFTs until enough samples have arrived"
Revert "Remove unnecessary clamping checks"
Revert "Shorten band names to fit in narrow spaces"
Revert "Add back clamping in band plan coordinate calculation"
Revert "Fix inconsistent 1-2-5 scaling of amplitude axis"
Revert "Initialize d_fftWindowType to zero"
Revert "dockfft: zoom slider uses log scale with max 100000x"
Revert "Switch to float in CPlotter::draw"
Revert "plotter: use volk kernels for some calculations"
Revert "Fix a double promotion warning in setNewFftData"
Revert "Fix some more double promotion warnings"
Revert "plotter: draw on pixel center for cleaner picture"
Revert "dockfft: add 333 fps rate for plotter"
Revert "plotter: raise rate limit, use different criteria for histogram calc"
Revert "plotter: limit plot update rate"
Revert "plotter: more responsive overlay redraws"
Revert "Calculate width of Y axis in saved waterfall"
Revert "Fix compilation"
Revert "Eliminate some implicit promotions to double"
Revert "Remove unnecessary function"
Revert "plotter: change "and" to "&&""
Revert "Store waterfall span setting in milliseconds"
Revert "Add fallback to read old fft_window setting"
Revert "dockfft: use strings for config file settings"
Revert "plotter: default markers off so they do not show on audio plotter"
Revert "dockfft: use original pandapter_ prefix in settings file"
Revert "plotter: use point instead of rectangle for wf drawing"
Revert "plotter: remove waterfall freeze feature"

+++ src/qtgui/dockbookmarks.cpp
@@ -158,20 +158,12 @@ void DockBookmarks::on_tableWidgetTagList_itemChanged(QTableWidgetItem *item)
 
 bool DockBookmarks::eventFilter(QObject* object, QEvent* event)
 {
-    // Since Key_Delete can be (is) used as a global shortcut, override the
-    // shortcut. Accepting a ShortcutOverride causes the event to be delivered
-    // again, but as a KeyPress.
-    if (event->type() == QEvent::KeyPress || event->type() == QEvent::ShortcutOverride)
+    if (event->type() == QEvent::KeyPress)
     {
         QKeyEvent* pKeyEvent = static_cast<QKeyEvent *>(event);
-        if (pKeyEvent->key() == Qt::Key_Delete)
+        if (pKeyEvent->key() == Qt::Key_Delete && ui->tableViewFrequencyList->hasFocus())
         {
-            if (event->type() == QEvent::ShortcutOverride) {
-                event->accept();
-            }
-            else {
-                return DeleteSelectedBookmark();
-            }
+            return DeleteSelectedBookmark();
         }
     }
     return QWidget::eventFilter(object, event);
